fix: rename enroll provider to provider enroll
Closes: #586
