+++ cmd/cli/main.go
@@ -21,12 +21,12 @@ import (
 	_ "github.com/stacklok/mediator/cmd/cli/app/apply"
 	_ "github.com/stacklok/mediator/cmd/cli/app/auth"
 	_ "github.com/stacklok/mediator/cmd/cli/app/docs"
-	_ "github.com/stacklok/mediator/cmd/cli/app/enroll"
 	_ "github.com/stacklok/mediator/cmd/cli/app/group"
 	_ "github.com/stacklok/mediator/cmd/cli/app/keys"
 	_ "github.com/stacklok/mediator/cmd/cli/app/org"
 	_ "github.com/stacklok/mediator/cmd/cli/app/policy"
 	_ "github.com/stacklok/mediator/cmd/cli/app/policy_status"
+	_ "github.com/stacklok/mediator/cmd/cli/app/provider"
 	_ "github.com/stacklok/mediator/cmd/cli/app/repo"
 	_ "github.com/stacklok/mediator/cmd/cli/app/role"
 	_ "github.com/stacklok/mediator/cmd/cli/app/rule_type"
